refactor(services): extract 4 leaf services into @flc/platform-services
Create shared @flc/platform-services package and move byte-identical leaf
services (loggingService, notificationService, businessReportService,
ticketAttachmentService) plus their tests into it. App-local copies become
re-export shims preserving the exact public surface (value/type exports split
for isolatedModules). Package code imports @flc/supabase directly; relocated
tests rewrite mocks to @flc/supabase. No behavior change.

Wire alias in root tsconfig.app/vite/vitest and hrms-web tsconfig/vite, add
workspace dep in both package.json. authService deferred (test divergence).

Validated: tsc clean, rpc/workflow checks pass, 996 tests pass, lint 0 errors,
root + hrms-web builds succeed.
