+++ apps/hrms-web/package.json
@@ -13,6 +13,7 @@
     "@flc/auth": "*",
     "@flc/hrms-schemas": "*",
     "@flc/hrms-services": "*",
+    "@flc/platform-services": "*",
     "@flc/shell": "*",
     "@flc/supabase": "*",
     "@flc/types": "*",
