+++ packages/platform-services/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@flc/platform-services",
+  "version": "0.0.1",
+  "private": true,
+  "description": "Shared cross-cutting platform services (logging, notifications, reporting, ticket attachments) for FLC BI apps",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  }
+}
