+++ packages/platform-services/src/index.ts
@@ -1,0 +1,5 @@
+// Barrel export for @flc/platform-services — shared cross-cutting services.
+export * from './loggingService';
+export * from './notificationService';
+export * from './businessReportService';
+export * from './ticketAttachmentService';
