+++ tsconfig.app.json
@@ -29,6 +29,7 @@
       "@flc/supabase": ["./packages/supabase/src/index.ts"],
       "@flc/shell": ["./packages/shell/src/index.ts"],
       "@flc/hrms-schemas": ["./packages/hrms-schemas/src/index.ts"],
+      "@flc/platform-services": ["./packages/platform-services/src/index.ts"],
       "@flc/ui/*": ["./packages/ui/src/*"]
     },
     "skipLibCheck": true,
